Revised example data.
Revised the example data to be more conservative. This is not meant as any kind of investing advice (this data is to show off the tool, not guide user asset allocations). Instead, the goal is to be more relatable to the typical buy-and-hold investor who would likely find the site valuable. Also made the example data more precise, to show off the calculator's floating point support.